Fixed the prototype in the file

+++ bit_manipulation/2-get_bit.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include "main.h"
 
-int get_bit(unsigned int n, unsigned int index)
+int get_bit(unsigned long int n, unsigned int index)
 {
     unsigned int mask = 1UL << index;
